Dreaming Throne rework
