+++ src/main/java/ti4/helpers/thundersedge/TeHelperBreakthroughs.java
@@ -7,7 +7,7 @@ import lombok.experimental.UtilityClass;
 import net.dv8tion.jda.api.components.buttons.Button;
 import net.dv8tion.jda.api.entities.channel.concrete.ThreadChannel;
 import net.dv8tion.jda.api.events.interaction.GenericInteractionCreateEvent;
-import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.DreamButtonHandler;
+import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.dream.DreamBreakthroughHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.theodisi.Ardentia.ArdentiaBreakthroughHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.theodisi.Kairn.KairnBreakthroughHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.whispers.arvaxi.ArvaxiBreakthroughHandler;
@@ -51,7 +51,7 @@ public class TeHelperBreakthroughs {
             case "nokarbt" -> TeHelperActionCards.beginPirates(game, player, "resolveNokarBt", 0, false);
             case "dihmohnbt" -> DSHelperBreakthroughs.dihmohnBTExhaust(game, player);
             case "cheiranbt" -> DSHelperBreakthroughs.cheiranBTExhaust(game, player);
-            case "dreambt" -> DreamButtonHandler.postDreamBtMoveNexusButtons(event, game, player);
+            case "dreambt" -> DreamBreakthroughHandler.postDreamBtMoveNexusButtons(event, game, player);
             case "kairnbt" -> KairnBreakthroughHandler.postInitialButtons(game, player);
             case "ardentiabt" -> ArdentiaBreakthroughHandler.startSubjugate(event, game, player);
             default -> {
