+++ src/main/java/ti4/service/leader/CommanderUnlockCheckService.java
@@ -3,7 +3,6 @@ package ti4.service.leader;
 import java.util.List;
 import java.util.Map.Entry;
 import lombok.experimental.UtilityClass;
-import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.DreamButtonHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.crystellum.CrystellumLeadersHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.ta.TaAbilityHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.theodisi.LostLegaciesCommanderUnlockHandler;
@@ -285,8 +284,18 @@ public class CommanderUnlockCheckService {
             }
 
             // BEANS
-            case "dream" ->
-                shouldBeUnlocked = (DreamButtonHandler.getNexusTokenTiles(game).size() >= 3);
+            case "dream" -> {
+                int eligibleSystems = 0;
+                for (Tile tile : game.getTileMap().values()) {
+                    if (!tile.isNebula(game)
+                            && tile.isAnomaly(game, player)
+                            && FoWHelper.playerHasActualShipsInSystem(player, tile)) {
+                        eligibleSystems++;
+                    }
+                }
+
+                shouldBeUnlocked = eligibleSystems >= 2;
+            }
             case "ta" -> shouldBeUnlocked = (TaAbilityHandler.getControlledPlanetCountWithAnyDesign(player, game) >= 4);
             case "netrunners" ->
                 shouldBeUnlocked = (ButtonHelper.getNumberOfUnitsOnTheBoard(game, player, "pds", false) >= 4);
