+++ src/main/java/ti4/service/turn/StartTurnService.java
@@ -14,8 +14,8 @@ import org.apache.commons.lang3.function.Consumers;
 import ti4.discord.interactions.buttons.Buttons;
 import ti4.discord.interactions.buttons.handlers.actioncards.theodisi.BorrowedTimeLLButtonHandler;
 import ti4.discord.interactions.buttons.handlers.actioncards.theodisi.PoliticalMarriageLLButtonHandler;
-import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.DreamButtonHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.ashen.AshenUnitHandler;
+import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.dream.DreamBreakthroughHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.netrunners.NetrunnersPromissoryHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.beans.ta.TaAbilityHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.theodisi.Aeterna.AeternaAbilityHandler;
@@ -572,7 +572,7 @@ public class StartTurnService {
                     startButtons.add(strategicAction);
                 }
             }
-            if (player.hasReadyBreakthrough("dreambt") && DreamButtonHandler.hasDreamBtNexusMove(game, player)) {
+            if (player.hasReadyBreakthrough("dreambt") && DreamBreakthroughHandler.hasDreamBtNexusMove(game, player)) {
                 startButtons.add(Buttons.gray(
                         factionChecker + "componentActionRes_exhaustBT_dreambt",
                         "Exhaust Dream-Space Convergence",
