+++ src/main/resources/data/abilities/beans_abilities.json
@@ -3,7 +3,7 @@
         "id": "dream_nexus",
         "name": "Dream Nexus",
         "faction": "dream",
-        "permanentEffect": "You may treat systems that contain nexus tokens as adjacent to each other.",
+        "permanentEffect": "Systems that contain your nexus tokens are nebulas. Your ship's movement is not affected by these nebulas.",
         "source": "beans"
     },
     {
@@ -23,35 +23,43 @@
         "source": "beans"
     },
     {
-        "id": "refraction",
-        "name": "Refraction",
-        "faction": "crystellum",
-        "window": "After another player produces 1 or more hits against your units during space combat:",
-        "windowEffect": "You may assign 1 of those hits to 1 of your non-fighter ships in an adjacent system.",
-        "source": "beans"
-    },
-    {
         "id": "fragmentation",
         "name": "Fragmentation",
         "faction": "crystellum",
-        "window": "After one or more of your non-fighter ships in the active system is destroyed during combat:",
-        "windowEffect": "You may place 1 fighter from your reinforcements in that system.",
+        "window": "After one of your non-fighter ships is destroyed:",
+        "windowEffect": "You may capture 1 fighter from the supply.",
+        "source": "beans"
+    },
+    {
+        "id": "pieces_of_a_whole",
+        "name": "Pieces of a Whole",
+        "faction": "crystellum",
+        "window": "At the start of space combat:",
+        "windowEffect": "You mayu return up to 2 captured fighters from your faction sheet to place that many fighters from your reinforcements in the active system.",
+        "source": "beans"
+    },
+    {
+        "id": "burning_legion",
+        "name": "Burning Legion",
+        "faction": "ashen",
+        "window": "During setup:",
+        "windowEffect": "Place 12 infantry from the supply on your faction sheet. These units are captured.",
         "source": "beans"
     },
     {
         "id": "phoenix_rising",
         "name": "Phoenix Rising",
         "faction": "ashen",
-        "window": "When you roll for your infantry's text ability:",
-        "windowEffect": "On a result of 9 or 10, you may choose for that unit to not be destroyed. If you do; you cannot resolve that unit's text ability.",
+        "window": "At the start of a round of ground combat on a planet you control:",
+        "windowEffect": "You may place 1 captured infantry from your faction sheet on that planet.",
         "source": "beans"
     },
     {
-        "id": "beauty_in_destruction",
-        "name": "Beauty in Destruction",
+        "id": "forged_in_fire",
+        "name": "Forged in Fire",
         "faction": "ashen",
-        "window": "Once per action, when one of your units is destroyed during combat:",
-        "windowEffect": "You may spend 1 token from your strategy pool to gain that unit's corresponding unit upgrade technology.",
+        "window": "When your infantry are destroyed or removed:",
+        "windowEffect": "Capture them.",
         "source": "beans"
     },
     {
@@ -59,7 +67,7 @@
         "name": "Mechanized Military",
         "faction": "iron",
         "window": "During setup:",
-        "windowEffect": "Take the 2 additional Iron Tide mechs and place them in your reinforcements; You have 6 mechs.",
+        "windowEffect": "Take the 3 additional Iron Tide mechs and place them in your reinforcements; You have 7 mechs.",
         "source": "beans"
     },
     {
@@ -73,7 +81,7 @@
         "id": "iron_resilience",
         "name": "Iron Resilience",
         "faction": "iron",
-        "permanentEffect": "Your ground forces cannot be removed or destroyed by any game effects except when in combat or when exceeding capacity.",
+        "permanentEffect": "Your mechs roll an additional combat die while they are damaged.",
         "source": "beans"
     },
     {
@@ -93,27 +101,27 @@
         "source": "beans"
     },
     {
-        "id": "system_breach",
-        "name": "System Breach",
+        "id": "neural_instruments",
+        "name": "Neural Instruments",
         "faction": "netrunners",
-        "window": "When another player places a structure:",
-        "windowEffect": "You may place one of that player's control tokens on your faction sheet.",
+        "window": "When any other player researches or gains a technology, or exhausts a technology for its effect:",
+        "windowEffect": "Place 1 of that player's control tokens on your faction sheet. You may not have more than 1 of each player's control token on your faction sheet at a time.",
         "source": "beans"
     },
     {
-        "id": "control_network",
-        "name": "Control Network",
+        "id": "proxy_network",
+        "name": "Proxy Network",
         "faction": "netrunners",
-        "window": "At any time:",
-        "windowEffect": "Before a player whose control token is on your faction sheet rolls dice for SPACE CANNON, you may remove 1 of those tokens to apply -1 to those dice. While you are the active player, you may remove a player's control token from your faction sheet to use the PRODUCTION ability of a space dock controlled by that player that you are blockading as if it were your own to produce ships.",
+        "window": "At the start of your turn:",
+        "windowEffect": "You may remove 1 player's control token from your faction sheet to copy the effect of 1 of that player's non-faction technologies until the end of your turn.",
         "source": "beans"
     },
     {
-        "id": "ransomware",
-        "name": "Ransomware",
+        "id": "reverse_engineering",
+        "name": "Reverse Engineering",
         "faction": "netrunners",
-        "window": "During the status phase:",
-        "windowEffect": "A player whose control token is on your faction sheet may give you 1 trade good to remove 1 of their control tokens there.",
+        "window": "Once per round, at the end of your turn:",
+        "windowEffect": "You may spend a token from your strategy pool to to gain a copied technology.",
         "source": "beans"
     },
     {
@@ -133,10 +141,10 @@
         "source": "beans"
     },
     {
-        "id": "natau_decree",
-        "name": "Decree",
+        "id": "dark_power",
+        "name": "Dark Power",
         "faction": "natau",
-        "permanentEffect": "You gain the text abilities of a doctrine while it is in your play area.",
+        "permanentEffect": "Each time you start or end your movement in an anomaly, gain 1 commodity.",
         "source": "beans"
     },
     {
