+++ src/main/resources/data/technologies/beans_technologies.json
@@ -9,7 +9,7 @@
         "faction": "dream",
         "baseUpgrade": "dd2",
         "source": "beans",
-        "text": "At the end of a tactical action in this unit's system, you may place or move 1 nexus token into any system that contains your units."
+        "text": "At the end of a tactical action in this unit's system, you may place or move 1 nexus token into a non-home system that contains 1 or more of your ships and no other player's ships."
     },
     {
         "alias": "bedreamneg",
@@ -20,7 +20,7 @@
         "requirements": "BB",
         "faction": "dream",
         "source": "beans",
-        "text": "During your tactical actions, apply +1 to the move value of any ship that starts its movement in a system that contains 1 or more of oyur nexus tokens."
+        "text": "During your tactical actions, apply +1 to the move value of any ship that starts its movement in a system that contains 1 or more of your nexus tokens. You may treat nebulas as adjacent to each other for the purposes of movement."
     },
     {
         "alias": "becrystda",
